relocation of some configs
